backend/pyproject.toml
[project]
name = "mini-app-backend"
version = "0.1.0"
description = "Backend for mini-app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.2",
    "fastapi[standard]>=0.119.0",
    "tortoise-orm>=0.25.1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
]
